Material bindings and descriptor sets now functional.

+++ Source/Include/Mesh.h
@@ -16,7 +16,7 @@ public:
 
     inline const uint64_t&   GetResourceHandle() const { return m_ResourceHandle; }
     inline const GfMatrix4f& GetLocalToWorld() const { return m_LocalToWorld; }
-    inline const uint32_t&   GetMaterialHash() const { return m_MaterialHash; }
+    inline const uint64_t&   GetMaterialHash() const { return m_MaterialHash; }
 
 protected:
 
@@ -33,7 +33,7 @@ private:
 
     GfMatrix4f m_LocalToWorld {};
 
-    uint32_t m_MaterialHash;
+    uint64_t m_MaterialHash;
 };
 
 #endif
